Add files via upload
This Python script is a web scraper designed to collect data about books available on the website "http://books.toscrape.com." It extracts book titles, image URLs, star ratings, prices, and availability information and stores it in a CSV file. The data can be used for various purposes, including analysis, reporting, or creating a catalog of books.